backend/requirements.txt
asgiref==3.5.2
asttokens==2.0.8
backcall==0.2.0
boto3==1.26.25
botocore==1.29.25
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.5
config==0.5.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
cymem==2.0.7
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
dj-rest-auth==2.2.5
Django==3.2.13
django-allauth==0.51.0
django-appconf==1.0.5
django-bootstrap5==22.1
django-cors-headers==3.13.0
django-environ==0.9.0
django-extensions==3.2.1
django-filter==22.1
django-imagekit==4.1.0
django-shortcuts==1.6
django-storages==1.13.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
executing==1.0.0
idna==3.4
inflection==0.5.1
ipython==8.5.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mypy-extensions==0.4.3
oauthlib==3.2.2
packaging==22.0
parso==0.8.3
pickleshare==0.7.5
pilkit==2.0
Pillow==9.2.0
prompt-toolkit==3.0.31
psycopg2-binary==2.9.5
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.2
Pygments==2.13.0
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.6
pyzmq==24.0.1
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
s3transfer==0.6.0
six==1.16.0
smart-open==5.2.1
spacy-legacy==3.0.10
spacy-loggers==1.0.3
sqlparse==0.4.3
stack-data==0.5.1
tomli==2.0.1
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
typing_extensions==4.4.0
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.12
wasabi==0.10.1
wcwidth==0.2.5
